chipAgent/requirements.txt
# Google ADK - Agent Development Kit
google-adk>=1.0.0

# Google GenAI SDK (dependency of ADK)
google-genai>=1.0.0

# Standard library dependencies (usually pre-installed)
# asyncio - built-in
# pathlib - built-in
# typing - built-in
# subprocess - built-in
# shutil - built-in
